Refactor dummy Sitter
